// src/core/plugins/series/bar-x/types.ts
import type {HtmlItem, LabelData, TooltipDataChunkBarX} from '../../../../types';
import type {PreparedAnnotation, PreparedBarXSeries} from '../../../series/types';

export type PreparedBarXData = Omit<TooltipDataChunkBarX, 'series'> & {
    type: 'bar-x';
    annotation?: PreparedAnnotation;
    x: number;
    y: number;
    width: number;
    height: number;
    opacity: number | null;
    series: PreparedBarXSeries;
    svgLabels: LabelData[];
    htmlLabels: HtmlItem[];
    isLastStackItem: boolean;
    /**
     * the utility field for storing the original height (for recalculations, etc.)
     * should not be used for displaying
     */
    _height: number;
};